Extract shared test types to common module
  Move 14 duplicated type definitions (Container, Config, Checkpoint, BlockHeader,
  Validator, TestState, Block, BlockBody, AggregatedAttestation, AggregationBits,
  AttestationData, ProposerAttestation, TestInfo, deser_pubkey_hex) from types.rs
  and signature_types.rs into a shared common.rs module. This eliminates the
  maintenance burden and risk of divergence between fork choice and signature tests.

  Also add autotests = false to Cargo.toml to prevent the helper modules from being
  auto-discovered as standalone test binaries.
