+++ crates/blockchain/Cargo.toml
@@ -8,6 +8,7 @@ readme.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 version.workspace = true
+autotests = false
 
 [features]
 # To skip signature verification during tests
